Atualizacoes 05/04/2015

+++ Classes/modelo/negocio/DadosEconMensais.java
@@ -15,7 +15,7 @@ public class DadosEconMensais {
     private int mes;
     private int ano;
     private String campo; // Nome do campo do tabela
-    private int quantidade;
+    private double quantidade;
     private double valorUnitario;
     private int tipoCampo; // A qual tabela o campo pertence? 1 - Entradas /2 - Saidas /3 - Despesas com Volumoso /4 - Concentrado /5 - Mineralização /6 - Medicamentos /7 - Ordenha /8 - Inseminação Artificial /9 - Despesas de Investimentos /10 - Despesas de Empréstimos /11 - COE da atividade leiteira
     private Perfil perfil;
@@ -23,7 +23,7 @@ public class DadosEconMensais {
     public DadosEconMensais() {
     }
 
-    public DadosEconMensais(int mes, int ano, String campo, int quantidade, double valorUnitario, int tipoCampo, Perfil perfil) {
+    public DadosEconMensais(int mes, int ano, String campo, double quantidade, double valorUnitario, int tipoCampo, Perfil perfil) {
         this.mes = mes;
         this.ano = ano;
         this.campo = campo;
@@ -65,11 +65,11 @@ public class DadosEconMensais {
         this.campo = campo;
     }
 
-    public int getQuantidade() {
+    public double getQuantidade() {
         return quantidade;
     }
 
-    public void setQuantidade(int quantidade) {
+    public void setQuantidade(double quantidade) {
         this.quantidade = quantidade;
     }
